misc changes on styling and took out maps component
