+++ src/styles/_app.scss
@@ -28,27 +28,6 @@ nav {
     background-color: $card-head-color !important    
 }
 
-/* width */
-::-webkit-scrollbar {
-    border-color: darkgrey;
-    border-radius: $scrollbar-radius;
-  }
-  
-  /* Track */
-  ::-webkit-scrollbar-track {
-    background-color: white;
-    border-color: darkgrey;
-    border-radius: $scrollbar-radius;
-  }
-  
-  /* Handle */
-  ::-webkit-scrollbar-thumb {
-    background-color: $scrollbar-handle-color;
-    border-color: darkgrey;
-    border-radius: $scrollbar-radius;
-  }  
- 
-
 // override twitter bootstrap
 @media (min-width: 57px) {
     .container {
