+++ src/styles/_list.scss
@@ -66,14 +66,12 @@ $restaurant-font-size: 0.8em;
     margin-left: -15px;
 }
 
-.restaurant-list {
+.restaurant-list-container {
     border-color: $border-color;
     border-style: groove;
-    border-radius: $border-radius;
     border-width: 2px;
     height: 1400px;
-    overflow: scroll;
-    padding: 1px;
+    
 
     li {
         padding:0 8px;
@@ -81,6 +79,9 @@ $restaurant-font-size: 0.8em;
     }
 }
 
+.restaurant-list {
+    overflow: scroll;    
+}
 
 .restaurant-list-label  {
     margin-bottom: 0  !important;
