+++ src/styles/map.scss
@@ -3,6 +3,10 @@
     height: 700px;
 }
 
+.restaurant-directions {
+    margin-top: -10px;
+}
+
 /* Phone-size screens */
 @media only screen and (max-width: 768px) {
     .map {
